[JENKNS-37425] continue refining Toast behavior; make reusable function for creating "Run Started" toast
