+++ blueocean-core-js/src/js/components/ReplayButton.jsx
@@ -6,9 +6,7 @@ import React, { Component, PropTypes } from 'react';
 import { Icon } from 'react-material-icons-blue';
 
 import { RunApi as runApi } from '../';
-import { ToastService as toastService } from '../';
-import { CAPABILITIES, capabilityStore } from '../';
-import { isMultiBranchRun, buildRunDetailsUrlFromQueue } from '../UrlBuilder';
+import { createRunStartedToast } from './utils';
 
 const stopProp = (event) => {
     event.stopPropagation();
@@ -37,55 +35,7 @@ export class ReplayButton extends Component {
         });
 
         runApi.replayRun(this.props.latestRun)
-            .then((runInfo) => this._nextAction(runInfo));
-    }
-
-    _nextAction(runInfo) {
-        const className = this.props.runnable._class;
-        const runId = runInfo.expectedBuildNumber;
-
-        const runDetailsUrl = buildRunDetailsUrlFromQueue(
-            runInfo._links.self.href,
-            false,
-            runId
-        );
-
-        this.props.onNavigation(runDetailsUrl);
-
-        /*
-        capabilityStore.resolveCapabilities(className)
-            .then(capabilityMap => {
-                const capabilities = capabilityMap[className];
-                const isMultiBranch = capabilities.some(capability => {
-                    return [CAPABILITIES.MULTIBRANCH_BRANCH, CAPABILITIES.MULTIBRANCH_PIPELINE].indexOf(capability) !== -1;
-                });
-
-                const runId = runInfo.expectedBuildNumber;
-                // TODO: href doesn't encode branch name correctly; verify bug fix after JENKINS-37873 is resolved
-                const runDetailsUrl = buildRunDetailsUrlFromQueue(
-                    runInfo._links.self.href,
-                    isMultiBranch,
-                    runId,
-                    true
-                );
-
-                if (this.props.autoNavigate && this.props.onNavigation) {
-                    this.props.onNavigation(runDetailsUrl);
-                } else {
-                    const name = decodeURIComponent(this.props.runnable.name);
-
-                    toastService.newToast({
-                        text: `Started "${name}" #${runId}`,
-                        action: 'Open',
-                        onActionClick: () => {
-                            if (this.props.onNavigation) {
-                                this.props.onNavigation(runDetailsUrl);
-                            }
-                        },
-                    });
-                }
-            });
-            */
+            .then((runInfo) => createRunStartedToast(this.props.runnable, runInfo, this.props.onNavigation));
     }
 
     render() {
