+++ blueocean-dashboard/package.json
@@ -35,7 +35,7 @@
     "skin-deep": "^0.16.0"
   },
   "dependencies": {
-    "@jenkins-cd/blueocean-core-js": "0.0.6-beta1",
+    "@jenkins-cd/blueocean-core-js": "0.0.7-beta1",
     "@jenkins-cd/design-language": "0.0.70",
     "@jenkins-cd/js-extensions": "0.0.22",
     "@jenkins-cd/js-modules": "0.0.6",
