+++ blueocean-personalization/package.json
@@ -34,7 +34,7 @@
     "react-addons-test-utils": "15.0.1"
   },
   "dependencies": {
-    "@jenkins-cd/blueocean-core-js": "0.0.6-beta1",
+    "@jenkins-cd/blueocean-core-js": "0.0.7-beta1",
     "@jenkins-cd/design-language": "0.0.70",
     "@jenkins-cd/js-extensions": "0.0.22",
     "@jenkins-cd/js-modules": "0.0.6",
